Refactor module structure to include 'backend' in import paths
- Updated the Go module path in go.mod from `github.com/OpenNSW/nsw` to `github.com/OpenNSW/nsw/backend`.
- Modified import paths across multiple files to reflect the new module structure, ensuring all internal packages are correctly referenced under the `backend` directory.
- Adjusted Makefile to use the updated import path for `goimports`.
- Ensured all affected files in the backend directory are updated to maintain functionality and consistency.
