+++ backend/internal/app/bootstrap/app.go
@@ -7,26 +7,26 @@ import (
 	"fmt"
 	"net/http"
 
-	"github.com/OpenNSW/nsw/internal/auth"
-	"github.com/OpenNSW/nsw/internal/config"
-	"github.com/OpenNSW/nsw/internal/consignment"
-	"github.com/OpenNSW/nsw/internal/database"
-	"github.com/OpenNSW/nsw/internal/hscode"
-	"github.com/OpenNSW/nsw/internal/middleware"
-	"github.com/OpenNSW/nsw/internal/payments"
-	"github.com/OpenNSW/nsw/internal/profile/cha"
-	"github.com/OpenNSW/nsw/internal/profile/company"
-	"github.com/OpenNSW/nsw/internal/profile/user"
-	taskmanager "github.com/OpenNSW/nsw/internal/task/manager"
-	"github.com/OpenNSW/nsw/internal/task/plugin"
-	"github.com/OpenNSW/nsw/internal/temporal"
-	workflowruntime "github.com/OpenNSW/nsw/internal/workflow/runtime"
-	"github.com/OpenNSW/nsw/internal/workflow/service"
-	"github.com/OpenNSW/nsw/pkg/storage"
-	"github.com/OpenNSW/nsw/pkg/storage/drivers"
+	"github.com/OpenNSW/nsw/backend/internal/auth"
+	"github.com/OpenNSW/nsw/backend/internal/config"
+	"github.com/OpenNSW/nsw/backend/internal/consignment"
+	"github.com/OpenNSW/nsw/backend/internal/database"
+	"github.com/OpenNSW/nsw/backend/internal/hscode"
+	"github.com/OpenNSW/nsw/backend/internal/middleware"
+	"github.com/OpenNSW/nsw/backend/internal/payments"
+	"github.com/OpenNSW/nsw/backend/internal/profile/cha"
+	"github.com/OpenNSW/nsw/backend/internal/profile/company"
+	"github.com/OpenNSW/nsw/backend/internal/profile/user"
+	taskmanager "github.com/OpenNSW/nsw/backend/internal/task/manager"
+	"github.com/OpenNSW/nsw/backend/internal/task/plugin"
+	"github.com/OpenNSW/nsw/backend/internal/temporal"
+	workflowruntime "github.com/OpenNSW/nsw/backend/internal/workflow/runtime"
+	"github.com/OpenNSW/nsw/backend/internal/workflow/service"
+	"github.com/OpenNSW/nsw/backend/pkg/storage"
+	"github.com/OpenNSW/nsw/backend/pkg/storage/drivers"
 
-	"github.com/OpenNSW/nsw/pkg/notification"
-	"github.com/OpenNSW/nsw/pkg/notification/channels"
+	"github.com/OpenNSW/nsw/backend/pkg/notification"
+	"github.com/OpenNSW/nsw/backend/pkg/notification/channels"
 )
 
 // App contains an initialized HTTP server and cleanup hooks.
