+++ backend/internal/auth/config.go
@@ -3,7 +3,7 @@ package auth
 import (
 	"fmt"
 
-	"github.com/OpenNSW/nsw/internal/validation"
+	"github.com/OpenNSW/nsw/backend/internal/validation"
 )
 
 type Config struct {
