+++ backend/internal/task/plugin/payment.go
@@ -7,7 +7,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/OpenNSW/nsw/internal/payments"
+	"github.com/OpenNSW/nsw/backend/internal/payments"
 	"github.com/google/uuid"
 	"github.com/shopspring/decimal"
 )
