+++ backend/internal/taskworkflow/persistence/store.go
@@ -7,7 +7,7 @@ import (
 
 	"gorm.io/gorm"
 
-	"github.com/OpenNSW/nsw/internal/task/plugin"
+	"github.com/OpenNSW/nsw/backend/internal/task/plugin"
 )
 
 // TaskWorkflowTask represents one persisted task within a task workflow.
