+++ backend/internal/workflow/service/template_service.go
@@ -5,7 +5,7 @@ import (
 
 	"gorm.io/gorm"
 
-	"github.com/OpenNSW/nsw/internal/workflow/model"
+	"github.com/OpenNSW/nsw/backend/internal/workflow/model"
 )
 
 type TemplateService struct {
