+++ backend/pkg/blobsource/config.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"time"
 
-	"github.com/OpenNSW/nsw/internal/validation"
+	"github.com/OpenNSW/nsw/backend/internal/validation"
 )
 
 // Config selects and configures a blob source backend.
