+++ backend/pkg/uiprojector/assembler_test.go
@@ -5,7 +5,7 @@ import (
 	"errors"
 	"testing"
 
-	"github.com/OpenNSW/nsw/pkg/uiprojector"
+	"github.com/OpenNSW/nsw/backend/pkg/uiprojector"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
