+++ backend/pkg/uiprojector/projector_test.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"testing"
 
-	"github.com/OpenNSW/nsw/pkg/uiprojector"
+	"github.com/OpenNSW/nsw/backend/pkg/uiprojector"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
