+++ backend/pkg/uiprojector/test/integration/integration_test.go
@@ -7,7 +7,7 @@ import (
 	"path/filepath"
 	"testing"
 
-	"github.com/OpenNSW/nsw/pkg/uiprojector"
+	"github.com/OpenNSW/nsw/backend/pkg/uiprojector"
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
